fix(shell): exhaustive caption dispatch, disabled-close captions, unified title-bar chromes
Address Greptile review:

- DesignCaptionButtons now dispatches explicitly per Windows variant and
  returns null for mac/win98/macos-glass instead of silently falling back
  to the Win11 glyphs for unknown variants.
- Add closeDisabled to CaptionButtonsProps so Windows-design Settings (and
  other non-closable surfaces) render dimmed, disabled caption close buttons
  instead of cross-design mac traffic lights; Settings uses caption buttons
  whenever the active design calls for them.
- Move the default mac pill chrome and the win98 raised chrome out of
  CanvasWindow into DesignTitleBarChrome.tsx (MacTitleBarChrome /
  Win98TitleBarChrome) so all five title-bar variants live in one module;
  DOM output is unchanged. 1357 shell tests pass, including new
  DesignCaptionButtons dispatch/disabled-close cases.
